+++ server/api/types/progress.go
@@ -3,5 +3,5 @@ package types
 type UpdateProgressRequest struct {
 	Manga   string `json:"manga"`
 	Chapter string `json:"chapter"`
-	Page    uint   `json:"page"`
+	Page    uint64 `json:"page"`
 }
