+++ server/db/models.go
@@ -3,7 +3,7 @@ package db
 import "time"
 
 type User struct {
-	Id        uint   `gorm:"primary_key"`
+	Id        uint64 `gorm:"primary_key"`
 	FirstName string `gorm:"not null"`
 	LastName  string
 	UserName  string
@@ -11,25 +11,25 @@ type User struct {
 }
 
 type Library struct {
-	Id         uint      `gorm:"primary_key"`
-	UserId     uint      `gorm:"index:idx_lib_user_manga,unique;not null"`
+	Id         uint64    `gorm:"primary_key"`
+	UserId     uint64    `gorm:"index:idx_lib_user_manga,unique;not null"`
 	Manga      string    `gorm:"index:idx_lib_user_manga,unique;not null"`
 	CoverImage string    `gorm:"not null"`
 	AddedAt    time.Time `gorm:"default:now()"`
 }
 
 type History struct {
-	Id     uint      `gorm:"primary_key"`
-	UserId uint      `gorm:"index:idx_hs_user_manga,unique;not null"`
+	Id     uint64    `gorm:"primary_key"`
+	UserId uint64    `gorm:"index:idx_hs_user_manga,unique;not null"`
 	Manga  string    `gorm:"index:idx_hs_user_manga,unique;not null"`
 	ReadAt time.Time `gorm:"default:now()"`
 }
 
 type Progress struct {
-	Id       uint      `gorm:"primary_key"`
-	UserId   uint      `gorm:"index:idx_prog_user_manga;not null"`
+	Id       uint64    `gorm:"primary_key"`
+	UserId   uint64    `gorm:"index:idx_prog_user_manga;not null"`
 	Manga    string    `gorm:"index:idx_prog_user_manga;not null"`
 	Chapter  string    `gorm:"index:idx_prog_user_manga;not null"`
-	Page     uint      `gorm:"not null;default:1"`
+	Page     uint64    `gorm:"not null;default:1"`
 	UpdateAt time.Time `gorm:"default:now()"`
 }
